+++ src/controllers/console/base/traits/GenerateClientsTableTrait.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace rhertogh\Yii2Oauth2Server\controllers\console\base\traits;
+
+trait GenerateClientTableTrait
+{
+
+}