+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Please check the [Upgrading Instructions](UPGRADE.md) when upgrading to a newer 
 
 ### Added
 - Support for default values in `Oauth2GeneratePatAction`. (rhertogh)
+- The `Oauth2ClientInterface` now defines `get`- and `set`-`MinimumSecretLength`.  (rhertogh)
 
 ### Changed